// teams/things/assets-catalog/base/src/domains/assets-v1/core/use-cases/create-asset/input-dto.schema.types.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ICreateAssetV1InputDto {
  /**
   * The name of the asset. It must have at least 1 character.
   */
  name: string;
  /**
   * The slug of the taxonomic unit the asset belongs to as a direct child. It must have at least 4 characters.
   */
  taxonomicUnitSlug: string;
  /**
   * An array of strings that help to describe the asset
   */
  tags: string[];
}